requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
pyjwt==2.8.0
redis==5.0.1
pynacl==1.5.0
base58==2.1.1
prometheus-client
httpx==0.26.0